Сделал фабрику и отрефакторил стартовые экраны таббара
